+++ test/e2e/framework/manifest/BUILD
@@ -8,7 +8,7 @@ load(
 go_library(
     name = "go_default_library",
     srcs = ["manifest.go"],
-    importpath = "k8s.io/kubernetes/test/e2e/manifest",
+    importpath = "k8s.io/kubernetes/test/e2e/framework/manifest",
     deps = [
         "//staging/src/k8s.io/api/apps/v1:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
