+++ test/e2e/framework/manifest/manifest.go
