+++ test/e2e/framework/statefulset/rest.go
@@ -33,7 +33,7 @@ import (
 	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/kubectl/pkg/util/podutils"
 	"k8s.io/kubernetes/test/e2e/framework"
-	"k8s.io/kubernetes/test/e2e/manifest"
+	e2emanifest "k8s.io/kubernetes/test/e2e/framework/manifest"
 )
 
 // CreateStatefulSet creates a StatefulSet from the manifest at manifestPath in the Namespace ns using kubectl create.
@@ -43,10 +43,10 @@ func CreateStatefulSet(c clientset.Interface, manifestPath, ns string) *appsv1.S
 	}
 
 	framework.Logf("Parsing statefulset from %v", mkpath("statefulset.yaml"))
-	ss, err := manifest.StatefulSetFromManifest(mkpath("statefulset.yaml"), ns)
+	ss, err := e2emanifest.StatefulSetFromManifest(mkpath("statefulset.yaml"), ns)
 	framework.ExpectNoError(err)
 	framework.Logf("Parsing service from %v", mkpath("service.yaml"))
-	svc, err := manifest.SvcFromManifest(mkpath("service.yaml"))
+	svc, err := e2emanifest.SvcFromManifest(mkpath("service.yaml"))
 	framework.ExpectNoError(err)
 
 	framework.Logf(fmt.Sprintf("creating " + ss.Name + " service"))
